mise en place de fucntion updateTweets et utilisation
